fix: local types for components
