+++ packages/react/src/types.ts
@@ -1,15 +1,7 @@
-import { UnionToIntersection } from './transformationTypes/helpers';
-import { Background } from './transformationTypes/background';
-import { Effect } from './transformationTypes/effect';
-import { ImageFormat, VideoFormat } from './transformationTypes/format';
-import { Quality } from './transformationTypes/quality';
 import { HeightOption, Resize, WidthOption } from './transformationTypes/resize';
-import { Rotate } from './transformationTypes/rotate';
-import { RoundCorners } from './transformationTypes/roundCorners';
-import { Opacity } from './transformationTypes/opacity';
-import { RemoveBackground } from './transformationTypes/removeBackground';
+import { UnionToIntersection } from './transformationTypes/helpers';
 
-type ResizeProps =
+export type ResizeProps =
   | {
       height?: HeightOption;
       width?: WidthOption;
@@ -18,38 +10,10 @@ type ResizeProps =
       resize?: Resize;
     };
 
-export type ImageTransformationProps = {
-  quality?: Quality;
-  format?: ImageFormat;
-  removeBackground?: RemoveBackground;
-  effects?: Effect[];
-  background?: Background;
-  rotate?: Rotate;
-  roundCorners?: RoundCorners;
-  opacity?: Opacity;
-} & ResizeProps;
-
-export type VideoTransformationProps = {
-  quality?: Quality;
-  format?: VideoFormat;
-  height?: HeightOption;
-  width?: WidthOption;
-  // removeBackground?: boolean | 'fineEdges';
-  // effects?: Effect[];
-  background?: Background;
-  rotate?: Rotate;
-  roundCorners?: RoundCorners;
-  opacity?: Opacity;
-};
-
-export type ImageParseTransformationProps = Partial<UnionToIntersection<ImageTransformationProps>>;
-export type VideoParseTransformationProps = Partial<UnionToIntersection<VideoTransformationProps>>;
-
-type TransformationNameToParser<Props> = {
+export type TransformationNameToParser<Props> = {
   [Key in keyof Props]-?: (
     value: Required<Props>[Key]
   ) => string;
 };
 
-export type ImageTransformationNameToParser = TransformationNameToParser<ImageParseTransformationProps>;
-export type VideoTransformationNameToParser = TransformationNameToParser<VideoParseTransformationProps>;
+export type TransformationMap<Props> = TransformationNameToParser<Partial<UnionToIntersection<Props>>>
